backend/app/services/embedding_service.py
import os
import numpy as np
from google import genai
from dotenv import load_dotenv

load_dotenv()

client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

class EmbeddingService:

    @staticmethod
    def embed_text(text: str) -> list[float]:
        result = client.models.embed_content(
            model="gemini-embedding-001",
            contents=text
        )
        return result.embeddings[0].values
